src/s3_upload_lambda/requirements.txt
requests==2.26.0
boto3==1.18.41
pillow==11.1.0